+++ addOns/authhelper/CHANGELOG.md
@@ -4,6 +4,9 @@ All notable changes to this add-on will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## Unreleased
+### Added
+- Add configuration support for the wait time after Client Script Based Authentication.
+
 ### Changed
 - Warn when the recorded script used with Client Script Based Authentication does not launch a browser.